UI/UX: Article card styling improvements
- Changed article cards to light gray background (#f9fafb) matching footer
- Removed borders from article cards for cleaner appearance
- Removed purple hover effect, keeping purple only for expanded state
- Improved visual hierarchy and reduced UI clutter
