+++ Extension/manifest.json
@@ -11,13 +11,13 @@
         "<all_urls>"
     ],
     "icons": {
-        "16": "images/icon-16.png",
-        "32": "images/icon-32.png",
-        "48": "images/icon-48.png",
-        "128": "images/icon-128.png"
+        "16": "assets/images/16x16.png",
+        "32": "assets/images/32x32.png",
+        "48": "assets/images/48x48.png",
+        "128": "assets/images/128x128.png"
     },
     "web_accessible_resources": [{
-        "resources": ["Readability.min.js"],
+        "resources": ["src/lib/Readability.min.js"],
         "matches": ["<all_urls>"]
     }]
 }
