+++ Extension/src/config/config.js
@@ -1,0 +1,4 @@
+export const CONFIG = {
+    apiKey: 'd4c9q1eNwP1ppeJvxU4wH2wczAjEX5wG8cGGz4C3',
+    apiUrl: 'https://gmoh9tv5t9.execute-api.eu-west-2.amazonaws.com/prod/vectormatch'
+};
