+++ Extension/styles/popup.css
@@ -39,6 +39,7 @@ body {
     padding: 16px;
     border-bottom: 1px solid #e5e7eb;
     flex-shrink: 0;
+    min-height: 56px;
 }
 
 .plan-tag {
@@ -87,6 +88,23 @@ body {
     font-size: 16px;
     font-weight: 600;
     color: #111827;
+}
+
+.logo-image {
+    height: 18px;
+    width: auto;
+    display: block;
+    object-fit: contain;
+    image-rendering: optimizeQuality;
+    shape-rendering: geometricPrecision;
+    will-change: transform;
+}
+
+.logo-icon-img {
+    height: 28px;
+    width: 28px;
+    display: block;
+    object-fit: contain;
 }
 
 .profile-avatar {
@@ -316,20 +334,19 @@ body {
     font-size: 13px;
     font-weight: 500;
     cursor: pointer;
-    transition: all 0.2s ease;
-    border-radius: 6px; 
+    transition: color 0.2s ease;
     position: relative;
+    border-bottom: 2px solid transparent;
 }
 
 .tab-button:hover {
     color: #9333ea;
-    background-color: #f3e8ff; 
+    background-color: transparent;
 }
 
 .tab-button.active {
     color: #9333ea;
-    background-color: #f3e8ff; 
-    font-weight: 500;
+    border-bottom-color: #9333ea;
 }
 
 .tab-button svg {
@@ -337,21 +354,6 @@ body {
     height: 16px;
 }
 
-.trending-badge {
-    position: absolute;
-    top: 4px;
-    right: 4px;
-    background-color: #dc2626;
-    color: white;
-    font-size: 10px;
-    font-weight: 600;
-    padding: 1px 5px;
-    border-radius: 10px;
-    min-width: 16px;
-    height: 16px;
-    line-height: 14px;
-    text-align: center;
-}
 
 .tab-content {
     flex: 1;
@@ -359,6 +361,7 @@ body {
     display: flex;
     flex-direction: column;
     min-height: 0;
+    max-height: 360px;
 }
 
 .tab-panel {
@@ -436,9 +439,9 @@ body {
 .articles-container {
     display: flex;
     flex-direction: column;
-    gap: 6px;
+    gap: 8px;
     overflow-y: auto;
-    padding: 4px 0;
+    padding: 8px;
     flex: 1;
     min-height: 0;
 }
@@ -446,9 +449,9 @@ body {
 .trending-container {
     display: flex;
     flex-direction: column;
-    gap: 8px;
+    gap: 0;
     overflow-y: auto;
-    padding: 4px 0;
+    padding: 0;
     flex: 1;
     min-height: 0;
 }
@@ -548,13 +551,10 @@ body {
 
 #treemap-container {
     width: 100%;
-    height: 300px;
-    border: 1px solid #e5e7eb;
-    border-radius: 8px;
-    background-color: #f9fafb;
+    height: 280px;
     position: relative;
     overflow: hidden;
-    margin-bottom: 12px;
+    margin: 0;
     animation: treeMapFadeIn 0.3s ease-out;
 }
 
@@ -674,9 +674,9 @@ body {
 }
 
 .article-card {
-    background-color: transparent;
-    border: 1px solid #e5e7eb;
-    border-radius: 8px;
+    background-color: #f9fafb;
+    border: none;
+    border-radius: 10px;
     padding: 0;
     transition: all 0.2s ease;
     cursor: pointer;
@@ -684,18 +684,78 @@ body {
     flex-shrink: 0;
 }
 
+.skeleton-article-card {
+    cursor: default;
+    pointer-events: none;
+}
+
+.skeleton-article-card.skeleton-processing {
+    background-color: #faf5ff;
+}
+
+.skeleton-stocks-container {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+}
+
+.skeleton-stock-item {
+    padding: 8px 10px;
+    background-color: rgba(255, 255, 255, 0.7);
+    border: none;
+    border-radius: 6px;
+}
+
+.loading-animation {
+    position: relative;
+    overflow: hidden;
+    pointer-events: none;
+}
+
+.loading-animation::after {
+    content: '';
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: linear-gradient(
+        90deg,
+        rgba(255, 255, 255, 0) 0%,
+        rgba(255, 255, 255, 0.3) 50%,
+        rgba(255, 255, 255, 0) 100%
+    );
+    animation: shimmer 1.5s infinite;
+}
+
+@keyframes shimmer {
+    0% {
+        transform: translateX(-100%);
+    }
+    100% {
+        transform: translateX(100%);
+    }
+}
+
+#treemap-container.loading-animation {
+    opacity: 0.6;
+}
+
+.article-card.loading-animation {
+    opacity: 0.6;
+}
+
 .article-card:hover {
-    border-color: #d1d5db;
     background-color: #f9fafb;
 }
 
 .article-card.expanded {
-    border-color: #9333ea;
     background-color: #faf5ff;
+    box-shadow: 0 2px 8px rgba(147, 51, 234, 0.15);
 }
 
 .article-collapsed {
-    padding: 10px 12px;
+    padding: 12px 14px;
 }
 
 .article-header {
@@ -708,17 +768,20 @@ body {
 .article-title {
     font-size: 13px;
     font-weight: 500;
-    color: #111827;
-    line-height: 1.3;
+    color: #1F2937;
+    line-height: 1.4;
     overflow: hidden;
     text-overflow: ellipsis;
-    white-space: nowrap;
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+    white-space: normal;
     flex: 1;
     margin-right: 8px;
 }
 
 .expand-indicator {
-    color: #6b7280;
+    color: #9CA3AF;
     transition: transform 0.2s ease, color 0.2s ease;
     flex-shrink: 0;
 }
@@ -731,11 +794,12 @@ body {
 .article-date {
     font-size: 11px;
     font-weight: 400;
-    color: #6b7280;
+    color: #9CA3AF;
+    margin-top: 4px;
 }
 
 .article-expanded {
-    border-top: 1px solid #e5e7eb;
+    border-top: 1px solid rgba(0, 0, 0, 0.06);
     padding: 10px 12px;
     max-height: 250px;
     overflow-y: auto;
@@ -743,8 +807,9 @@ body {
 }
 
 .article-card.expanded .article-expanded {
-    border-top-color: #e9d5ff;
+    border-top-color: rgba(147, 51, 234, 0.1);
 }
+
 
 @keyframes expandIn {
     from {
@@ -768,18 +833,18 @@ body {
 }
 
 .stock-item {
-    background-color: #f9fafb;
-    border: 1px solid #e5e7eb;
+    background-color: rgba(255, 255, 255, 0.7);
+    border: none;
     border-radius: 6px;
-    padding: 6px 8px;
+    padding: 8px 10px;
     cursor: pointer;
     transition: all 0.2s ease;
 }
 
 .stock-item:hover {
-    border-color: #9333ea;
-    background-color: #f3e8ff;
+    background-color: rgba(255, 255, 255, 1);
     transform: translateX(2px);
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
 }
 
 .stock-header {
